requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
InstagramGrabber==1.0.6
python-dotenv==1.0.0
requests==2.31.0